Classes encapsulées + controlleur classemennt
