--- app/src/main/java/com/example/footballapi/model/competition/Competition.java
@@ -1,19 +1,0 @@
-package com.example.footballapi.model.competition;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Competition {
-
-    @SerializedName("name")
-    @Expose
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
