+++ app/src/main/res/layout/fragment_list_classement.xml
@@ -6,19 +6,10 @@
     android:orientation="vertical"
     tools:context="com.example.footballapi.view.ClassementFragment">
 
-    <TextView
-        android:id="@+id/tvClassement"
-        android:layout_width="match_parent"
-        android:layout_height="80dp"
-        android:textColor="@color/black"
-        android:paddingLeft="20dp"
-        android:textSize="30dp"
-        android:layout_weight="1" />
-
-    <ListView
+        <ListView
         android:id="@+id/lvClassement"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
         android:layout_weight="1">
     </ListView>
